requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpaca-trade-api==0.48
alpha_vantage==3.0.0
asyncio-nats-client==0.11.5
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
curl_cffi==0.13.0
Deprecated==1.2.18
frozendict==2.4.6
frozenlist==1.7.0
idna==3.10
lxml==6.0.2
multidict==6.6.4
multitasking==0.0.12
numpy==2.3.3
pandas==2.3.2
peewee==3.18.2
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.1
pycparser==2.23
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
schedule==1.2.2
sec-edgar-downloader==5.0.3
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.24.3
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
yfinance==0.2.66
